requirements.txt
feedparser==6.0.11
requests==2.32.4
beautifulsoup4==4.12.3
Pillow==10.3.0
python-dotenv==1.0.1
platformdirs==4.2.0
PyYAML==6.0.1
ImageHash==4.3.1
telethon>=1.36.0
